next gen

+++ include/mocap/utils.hpp
@@ -1,4 +1,4 @@
 #include "opencv2/core/matx.hpp"
 #include "tf2/LinearMath/Quaternion.h"
 
-void quaternionFromRvecs(const cv::Vec3f &rvec, tf2::Quaternion &q);
+void quaternionFromRvecs(const cv::Vec3d &rvec, tf2::Quaternion &q);
